fix: add kubernetes metadata correctly to payload (#56)
Previously the Kubernetes metadata was added to `meatadata.kubernetes`, but the APM Server expects it to be under `metadata.system.kubernetes`, so it was just silently dropped.

I took the liberty to limit the creation of hidden classes a bit while I was touching the code. It's not a hot code path, so it's not a big deal, but I thought I'd do it anyway.