test: allow test/side-effects.js to be run stand-alone (#87)
If running `tape test/*.js`, the tests inside `test/side-effects.js`
would run just fine. But when running the file directly, either via
`tape test/side-effects.js` or `node test/side-effects.js`, the tests
would fail because of a race-condition.

The race-condition meant that the order of the callback calls would be
different. By omitting the `t.end()` call and relying on the `t.plan()`
call instead, we get around this issue.