feat: add serverCaCert option (#102)
We add support for specifying the CA certificate
for verifying the server's TLS certificate chain.

Update the tests to use a self-signed certificate
that is valid for "localhost", rather than the
one generated by self-signed, which is valid for
"example.org". This certificate also has a much
longer expiry time, which avoids having to reinstall
http-pem after a year.