update client.flush() to signal a current intake req to finish quickly
To support a lambda function invocation, an agent.flush(), at least for
a sync lambda handler
(https://docs.aws.amazon.com/lambda/latest/dg/nodejs-handler.html#nodejs-handler-sync),
needs to socket.ref() so that the APM server communication is considered
an active handle so that function execution continues.

This also changes the APM server response timeout to the shorter value
so that a slow APM server won't hang the function invocation for as
long.